Add a naive class Player.
Class Player inherits from class Entity.
Encabulate some methods for class Player.
For example, get the maximum number of available bombs of the player, set bombs and so on.
Since class Bomb has not been finished yet, I just add a sprite to the GameScene whenever a bomb is setting.