Add a naive class Bomb
This is the first naive version for the Bomb class. The primary goal is to design a bomb which can explode in a certain time. I initially intended to use call back function with update function to finish the task. However, there was something wrong confusing me for a couple of days, and the first solution for me was to deal with the removal of Sprite. In fact, the mistake was where the application of call back function was used. After the removal of call back function, the Bomb can explode successfully. The further goal is to integrate explosion animation and the affect caused by explosion among entities into the current Bomb class.